refactor(basket): Phase 7 test reshape to Features/<Slice>/
- Split BasketApiEndpointsTests into 5 per-slice handler test files;
  moved MessagingProviderBootTests to Infrastructure/; deleted now-empty
  Endpoints/ and IntegrationEvents/ folders. Test count unchanged.
- Scoped each slice's MetricFactory test double to a per-slice meter
  name: split classes now run in parallel and shared "Basket.Tests"
  meter let one class's MeterListener observe another's measurements.
- No OrderCreated/ProductPriceUpdated test folders: no event-consumer
  handler tests exist to move (PRD adds no new feature tests).
